Modified "dti_distance_presetSigma.m" so that it computes a 2D histogram and filters it.
This version of the Matlab implementation is closer to the Atropos implementation of simply filtering a joint histogram with a 2D Gaussian filter given a preset sigma.  This replaces the "kde2d" function which calculates an optimal sigma based on characteristics of the data. This implementation requires the "hist2" function available at http://www.mathworks.com/matlabcentral/fileexchange/9896-2d-histogram-calculation.