Modified "getPrinEigVec.m" so that it doesn't flip the x component of the matrix.
We realized that Atropos returns eigenvectors that range throughout the entire sphere, but our Matlab code does not.  We realized that we had forced the x component of the eigenvector to be positive because we only wanted outputs on the positive x direction (one hemisphere), but because Atropos does not do that, we're taking that out.